Updated application sources resolvers
